# backend/openapi/v1.yaml
openapi: 3.0.3
info:
  title: DataAfrik API
  version: 1.0.0
  description: >
    API-first contract for DataAfrik domain services. This v1 spec is a migration
    baseline and is expanded incrementally while legacy endpoints remain active.
servers:
  - url: /api/v1
...
paths:
...
  /finance/ingestion/jobs:
    post:
      tags: [finance]
      summary: Queue finance ingestion job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rows]
              properties:
                rows:
                  type: array
                  items:
                    type: object
                mapping:
                  type: object
                options:
                  type: object
      responses:
        '202':
          description: Job queued
        '200':
          description: Existing idempotent job reused
  /finance/reconciliation/jobs:
    post:
      tags: [finance]
      summary: Queue bank reconciliation job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bankRows]
              properties:
                bankRows:
                  type: array
                  items:
                    type: object
                bankMapping:
                  type: object
                bookRows:
                  type: array
                  items:
                    type: object
                bookTransactions:
                  type: array
                  items:
                    type: object
                bookMapping:
                  type: object
                options:
                  type: object
      responses:
        '202':
          description: Job queued
        '200':
          description: Existing idempotent job reused
  /finance/reports/jobs:
    post:
      tags: [finance]
      summary: Queue financial summary report job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transactions:
                  type: array
                  items:
                    type: object
                rows:
                  type: array
                  items:
                    type: object
                mapping:
                  type: object
                options:
                  type: object
                companyName:
                  type: string
                reportPeriod:
                  type: string
      responses:
        '202':
          description: Job queued
        '200':
          description: Existing idempotent job reused
  /finance/jobs:
    get:
      tags: [finance]
      summary: List authenticated user's finance jobs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: Job list
  /finance/jobs/{jobId}:
    get:
      tags: [finance]
      summary: Get finance job status/result
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Job details
        '404':
          description: Job not found
  /finance/jobs/storage-mode:
    get:
      tags: [finance]
      summary: Get active finance job storage mode
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Storage mode info
...
  /subscriptions/pdf-checkout:
    post:
      tags: [subscriptions]
      summary: Start PDF checkout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Checkout initialized
  /subscriptions/pdf-access:
    get:
      tags: [subscriptions]
      summary: Verify PDF entitlement
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Access state
  /subscriptions/pdf-verify:
    get:
      tags: [subscriptions]
      summary: Verify checkout session/reference
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Verification status
  /subscriptions/newsletter:
    post:
      tags: [subscriptions]
      summary: Subscribe email to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Subscription result
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...